peakfit: move compute to a separate module
Function called on individual process are moved to a separate module for
faster process initialization.
